fix: base archtect
