+++ cmd/clean/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"encoding/json"
+	"github.com/burhon94/clean/cmd/clean/router"
 	"log"
 	"net/http"
 )
@@ -12,7 +12,7 @@ func main() {
 		addr   = ":9999"
 	)
 
-	newMux := initRoute(prefix)
+	newMux := router.InitRouter(prefix)
 
 	newServer := http.Server{
 		Handler: newMux,
@@ -29,38 +29,4 @@ func main() {
 		newServer.ListenAndServe()
 
 	}()
-}
-
-func initRoute(prefix string) *http.ServeMux {
-
-	newServeMux := http.NewServeMux()
-
-	newServeMux.HandleFunc(prefix+"/ping", pingHandler)
-
-	return newServeMux
-}
-
-type responseStruct struct {
-	Code    int         `json:"code"`
-	Payload interface{} `json:"payload"`
-	Message string      `json:"message"`
-}
-
-func pingHandler(w http.ResponseWriter, r *http.Request) {
-	var resp = responseStruct{
-		Code:    202,
-		Payload: "pong",
-		Message: "Success",
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-
-
-	data, err := json.Marshal(resp)
-	if err != nil {
-		w.WriteHeader(500)
-		return
-	}
-
-	w.Write(data)
 }
