+++ cmd/clean/router/router.go
@@ -1,0 +1,17 @@
+package router
+
+import (
+	"github.com/burhon94/clean/cmd/clean/handlers"
+	"net/http"
+)
+
+func InitRouter(prefix string) *http.ServeMux {
+
+	newServeMux := http.NewServeMux()
+
+	newServeMux.HandleFunc(prefix+"/ping", handlers.HandlerPing)
+
+	return newServeMux
+}
+
+
